Alerting: add mute timings provenance to config api
